requirements.txt
Django==4.2
django_filter==23.1
djangorestframework==3.14.0
django-log-request-id
markdown